+++ server/app/src/dtos/groups/save-group.dto.ts
@@ -1,4 +1,4 @@
-export class CreateGroupDto {
+export class SaveGroupDto {
   day: string;
   hour: string;
   address: string;
