--- server/app/src/dtos/groups/update-group.dto.ts
@@ -1,11 +1,0 @@
-export class UpdateGroupDto {
-  id?: string;
-  day: string;
-  hour: string;
-  address: string;
-  room: string;
-  startDate: string;
-  endDate: string;
-  lecturers: string[];
-  students: string[];
-}
