+++ server/app/src/mappers/lecturer.mapper.ts
@@ -1,4 +1,4 @@
-import { SaveLecturerDto } from '../dtos/lecturers/create-lecturer.dto';
+import { SaveLecturerDto } from '../dtos/lecturers/save-lecturer.dto';
 import { Lecturer } from '../models/lecturer.model';
 import { LecturerDto } from '../dtos/lecturers/lecturer.dto';
 import { Group } from '../models/group.model';
